requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
aiosqlite==0.19.0
stripe==9.8.0
flask==3.0.0